Roll back to unsigned long long to interact with libsodium

+++ src/cipher.cpp
@@ -127,8 +127,8 @@ void Cipher::CipherImpl::encrypt(const unsigned char* in,
             "The minimum number of bytes to encrypt is 1.");
     }
 
-    uint8_t  chacha_key[crypto_aead_chacha20poly1305_KEYBYTES];
-    uint64_t c_len = 0;
+    uint8_t            chacha_key[crypto_aead_chacha20poly1305_KEYBYTES];
+    unsigned long long c_len = 0; // NOLINT
 
     // generate a random nonce, and place it at the beginning of the output
     random_bytes(NONCE_SIZE, out);
@@ -189,8 +189,8 @@ void Cipher::CipherImpl::decrypt(const unsigned char* in,
             + std::to_string(ciphertext_length(0))); /* LCOV_EXCL_LINE */
     }
 
-    uint8_t  chacha_key[crypto_aead_chacha20poly1305_KEYBYTES];
-    uint64_t m_len = 0;
+    uint8_t            chacha_key[crypto_aead_chacha20poly1305_KEYBYTES];
+    unsigned long long m_len = 0; // NOLINT
 
     // unlock the master key
     key_.unlock();
